Separated walker and snapshotter behavior for broken symlinks
